+++ plugins/de.cau.cs.kieler.klay.force/src/de/cau/cs/kieler/klay/force/model/FruchtermanReingoldModel.java
@@ -4,7 +4,7 @@
  * http://www.informatik.uni-kiel.de/rtsys/kieler/
  * 
  * Copyright 2011 by
- * + Christian-Albrechts-University of Kiel
+ * + Kiel University
  *   + Department of Computer Science
  *     + Real-Time and Embedded Systems Group
  * 
@@ -48,7 +48,7 @@ public final class FruchtermanReingoldModel extends AbstractForceModel {
     @Override
     protected void initialize(final FGraph graph) {
         super.initialize(graph);
-        temperature = graph.getProperty(Properties.TEMPERATURE);
+        temperature = graph.getProperty(Properties.TEMPERATURE).doubleValue();
         threshold = temperature / graph.getProperty(Properties.ITERATIONS);
         
         // calculate an appropriate value for K
